chore: refactor sdk types so that source/destination postgres share creds
