+++ apps/sample-app/supaglue-config/postgres_credentials.ts
@@ -3,7 +3,7 @@ import * as dotenv from 'dotenv';
 
 dotenv.config({ path: `${__dirname}/../.env` });
 
-const credentials = sdk.sources.postgresCredentials({
+const credentials = sdk.internalIntegrations.postgresCredentials({
   host: 'postgres',
   port: 5432,
   database: 'sample_app',
